01/03

+++ utils.c
@@ -1,6 +1,6 @@
 #include "push_swap.h"
 
-int	ft_atol(const char *str)
+long	ft_atol(const char *str)
 {
 	long			res;
 	int			signal;
